refactor: 將 StoryImage 移出 StoryContainer
